add like_button and article_progile widgets
